@@ -17,7 +17,6 @@ Gem::Specification.new do |gem|
   gem.test_files    = gem.files.grep(%r{^(test|spec|features)/})
   gem.require_paths = ["lib"]
 
-  gem.add_dependency("rails", "~> 4.0.0")
-  gem.add_dependency("friendly_id", "5.0.0.beta4")
-  gem.add_dependency("protected_attributes")
+  gem.add_dependency("rails", "~> 5.1")
+  gem.add_dependency("friendly_id")
 end
