@@ -1,4 +1,4 @@
-class CreateSongRefTable < ActiveRecord::Migration
+class CreateSongRefTable < ActiveRecord::Migration[5.0]
 	def up
 		create_table :song_refs, :id => false do |t|
 			t.string :uuid, :primary => true, :null => false
