@@ -1,7 +1,7 @@
 require 'yaml'
 require 'gdshowsdb'
 
-class ImportSongRefs < ActiveRecord::Migration
+class ImportSongRefs < ActiveRecord::Migration[5.0]
 	def up		
 		Gdshowsdb::SongRefYAMLParser.from_yaml.parse.each do |song_ref_hash|
 			SongRef.create_from(song_ref_hash)
