@@ -1,4 +1,4 @@
-class CreateSongTable < ActiveRecord::Migration
+class CreateSongTable < ActiveRecord::Migration[5.0]
 	
 	def up
 		create_table :songs, :id => false do |t|
