@@ -1,4 +1,4 @@
-class AddCountToSongRefTable < ActiveRecord::Migration
+class AddCountToSongRefTable < ActiveRecord::Migration[5.0]
 
   def up
     add_column :song_refs, :song_occurences_count, :integer, :default => 0
