@@ -5,5 +5,18 @@ class SongOccurence < ActiveRecord::Base
   belongs_to :song_ref, :foreign_key => :song_ref_uuid, :primary_key => :uuid, :counter_cache => true
   accepts_nested_attributes_for :show
 
-  attr_accessible :position, :uuid, :show
+  def self.create_from(spec)
+    occurence = SongOccurence.new
+    set_spec(occurence, spec)
+    occurence.save
+    occurence
+  end
+
+  private
+
+  def self.set_spec(occurence, spec)
+    occurence.uuid = spec[:uuid]
+    occurence.position = spec[:position]    
+    occurence.show = ShowSet.find_by_uuid(spec[:show_set_uuid]).show
+  end
 end
